adwawdawddawoijsdfkjsefnj
ejdabdawdajadjsmkasjksakskskskdkjsjdjsjsjsjsdnjdnfnc   jjejrj. ejejrjeorieie dorowo nanfnre.w fnjdajnfwejr ,.rkekekwkrkw x.AOQOSIDHEJRJ  The path of the righteous man is beset on all sides by the inadequacies of the selfish and the tyrany of evil men. Blessed is he who in the name of charity and good will sheperds the through the valley of darkness, for he is truly his brothers keeper and the finder of lost children. And i will strike down upon thee with great vengeance and furious anger those who attempt to poison and destroy my brothers. And you will know my name is The Lord when i lay my vengeance upon thee.